requirements.txt
rpi-ws281x==4.2.2
RPi.GPIO==0.7.0
websockets==8.1
